+++ tests/test_texture_array.py
@@ -33,3 +33,14 @@ class TestCase(unittest.TestCase):
 
         texture.build_mipmaps()
         assert texture.filter == (moderngl.LINEAR_MIPMAP_LINEAR, moderngl.LINEAR)
+
+    def test_texture_default_filter(self):
+        """Ensure default filter is correct"""
+        # Float types
+        for dtype in ["f1", "f2", "f4"]:
+            texture = self.ctx.texture_array((10, 10, 10), 4, dtype=dtype)
+            self.assertEqual(texture.filter, (moderngl.LINEAR, moderngl.LINEAR))
+
+        for dtype in ["u1", "u2", "u4", "i1", "i2", "i4"]:
+            texture = self.ctx.texture_array((10, 10, 10), 4, dtype=dtype)
+            self.assertEqual(texture.filter, (moderngl.NEAREST, moderngl.NEAREST))