+++ tests/test_texture_cube.py
@@ -67,6 +67,17 @@ class TestCase(unittest.TestCase):
         with self.assertRaises(Exception):
             tex.read(6)
 
+    def test_texture_default_filter(self):
+        """Ensure default filter is correct"""
+        # Float types
+        for dtype in ["f1", "f2", "f4"]:
+            texture = self.ctx.texture_cube((10, 10), 4, dtype=dtype)
+            self.assertEqual(texture.filter, (moderngl.LINEAR, moderngl.LINEAR))
+
+        for dtype in ["u1", "u2", "u4", "i1", "i2", "i4"]:
+            texture = self.ctx.texture_cube((10, 10), 4, dtype=dtype)
+            self.assertEqual(texture.filter, (moderngl.NEAREST, moderngl.NEAREST))
+
 
 if __name__ == '__main__':
     unittest.main()